p8fs/pyproject.toml
[project]
name = "p8fs"
version = "1.2.7"
description = "Core percolate memory system with RAG/IR features and database repositories"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "apscheduler>=3.10.4",
    "p8fs-cluster",
    "pydantic>=2.5.0",
    "markdown>=3.5.0",
    "psycopg2-binary>=2.9.0",
    "tenacity>=8.0.0",
    "typer>=0.12.0",
    "nats-py>=2.7.0",
    "pyyaml>=6.0.0",
    "pymysql>=1.1.0",
    "dbutils>=3.1.0",
    "semantic-text-splitter>=0.28.0",
    "requests>=2.32.3",
    "pillow>=11.0.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "urllib3>=2.2.3",
    "certifi>=2024.7.4",
    "setuptools>=75.3.0",
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
    "fastembed>=0.7.0",
    "tikv-client>=0.0.4",
    "tiktoken>=0.8.0",
]

[project.scripts]
p8fs = "p8fs.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]
chunking = [
    "semchunk>=3.2.3",
]
workers = [
    "p8fs-node",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.uv.sources]
p8fs-cluster = { workspace = true }
p8fs-node = { workspace = true }

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.15.0",
]
